Modify this package to use the functions in github.com/google/uuid where appropriate.  There should be no functional change with this commit.
This has been done to reduce the amount of redundant code between
the two packages, in particular, code dealing with the clock, random
numbers, and determining the node ID from the network interfaces.

Where reasonable, this package is just a wrapper around google/uuid.
There still is duplicated code to limit the performance impact of
wrapping small functions.

The Version, Variant and Time types are now aliases of the google/uuid types.